+++ app/Http/Controllers/TypeDocumentController.php
@@ -152,6 +152,13 @@ class TypeDocumentController extends Controller
             DB::beginTransaction();
             $typeVehicle = $this->typeDocumentRepository->find($id);
             if ($typeVehicle) {
+                
+                if (
+                    $typeVehicle->VehicleDocuments()->exists()
+                ) {
+                    throw new \Exception('No se puede eliminar el registro, por que tiene relación de datos en otros módulos');
+                }
+
                 $typeVehicle->delete();
                 $msg = 'Registro eliminado correctamente';
             } else {
@@ -165,7 +172,7 @@ class TypeDocumentController extends Controller
 
             return response()->json([
                 'code' => 500,
-                'message' => Constants::ERROR_MESSAGE_TRYCATCH,
+                'message' => $th->getMessage(),
                 'error' => $th->getMessage(),
                 'line' => $th->getLine(),
             ], 500);
